Moving validation function to component
